update stats after import
